Update setup-teslausb
Got the "mount error(95): Operation not supported" error when running setup-teslausb to mount a share on a server that does not support CIFS 3.0. Message to the user suggests this: "

Try specifying a lower number for the CIFS version like this: export cifs_version=2

This actually doesn't work. Changing this to:

Try specifying a lower number for the CIFS version like this: export cifs_version=2.0

This works and is more helpful to the user.